built the skeleton of the Contacts Fragment
Contacts Fragment shows a list of contacts the user has.  It also allows the user to add contact.  Firebase client will post a request to the target user.  The target user needs to accept the request before the original user can add him in the contact.  hooking up the Firebase Cloud Functions to  process the contact request with Firebase database.